only calculate changes in branch if necessary
